# tg-clicker/server/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Girand Clicker",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://clicker-backend.inlostcity.com/api/"
    }
  ],
  "paths": {
...
    "/guild/create": {
      "post": {
        "tags": [
          "guild"
        ],
        "summary": "Create guild",
        "description": "Create guild",
        "operationId": "createGuild",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/guildCreateReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/guildCreateRes"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "guildCreateReq": {
        "type": "object",
        "properties": {
          "guildId": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "owner": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          }
        }
      },
      "guildCreateRes": {
        "type": "object",
        "properties": {
          "refLink": {
            "type": "string"
          }
        }
      },
...
    }
  }
}